Bag_design class for common methods and parameters
Verified with resistor
